requirements.txt
django-acknowledge
django-audience>=0.5.3
django-categories>=1.0.2
django-cms
django-core-media
django-concepts>=0.3.5
django-credits
django-edumetadata>=0.6
django-licensing>=0.3,<1.0
django-localflavor>=1.0
django-ngs-taxonomy
django-quotations
django-reference
django-resourcecarousel>=1.0.1
django-treebeard
beautifulsoup4